+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Proper debug.traceback level
+
 ### Added
 
 - Implement `wrap()` function for postprocessing net.box call results